# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Price Guide API with comp-based valuation engine
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /items/{slug}/valuation:
    get:
      operationId: getItemValuation
      tags: [valuation]
      summary: Get valuation for an item
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Valuation result with audit trail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValuationResult"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{slug}/sales:
    get:
      operationId: getItemSales
      tags: [valuation]
      summary: Get sales history for an item (show your work)
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sales records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /items/{slug}/comps:
    get:
      operationId: getItemComps
      tags: [valuation]
      summary: Get comparable items used in valuation
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comparable items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompMatch"
...
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get current user's watchlist
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemSummary"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add item to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistEntry"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /watchlist/{itemId}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove item from watchlist
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    ItemSummary:
      type: object
      required: [id, slug, name, category, status, createdAt]
      properties:
        id:
          type: number
        slug:
          type: string
        name:
          type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        year:
          type: ["number", "null"]
        authenticator:
          type: ["string", "null"]
        notorietyTier:
          type: ["string", "null"]
        imageUrls:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        watchCount:
          type: number
        medianEstimate:
          type: ["number", "null"]
        confidence:
          type: ["string", "null"]
        createdAt:
          type: string
...
    SaleRecord:
      type: object
      required: [id, itemId, salePrice, saleDate, saleSource, verified, createdAt]
      properties:
        id:
          type: number
        itemId:
          type: number
        itemName:
          type: ["string", "null"]
        itemSlug:
          type: ["string", "null"]
        salePrice:
          type: number
        saleDate:
          type: string
        saleSource:
          type: string
          enum: [auction_house, private, marketplace, admin_estimate]
        sourceName:
          type: ["string", "null"]
        conditionGrade:
          type: ["string", "null"]
        authenticationStatus:
          type: ["string", "null"]
        verified:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    ValuationSegment:
      type: object
      required: [conditionGrade, authenticationStatus, sampleSize]
      properties:
        conditionGrade:
          type: ["string", "null"]
        authenticationStatus:
          type: ["string", "null"]
        sampleSize:
          type: number
        lowEstimate:
          type: ["number", "null"]
        medianEstimate:
          type: ["number", "null"]
        highEstimate:
          type: ["number", "null"]
        confidence:
          type: ["string", "null"]
        singleSaleNote:
          type: ["string", "null"]
        sales:
          type: array
          items:
            $ref: "#/components/schemas/SaleRecord"
    ValuationResult:
      type: object
      required: [itemId, tierUsed, segments, computedAt]
      properties:
        itemId:
          type: number
        itemSlug:
          type: string
        tierUsed:
          type: string
          enum: [A, B, C]
        segments:
          type: array
          items:
            $ref: "#/components/schemas/ValuationSegment"
        categoryRangeNote:
          type: ["string", "null"]
        categoryLow:
          type: ["number", "null"]
        categoryHigh:
          type: ["number", "null"]
        computedAt:
          type: string
    CompMatch:
      type: object
      required: [id, compItemId, similarityScore]
      properties:
        id:
          type: number
        compItemId:
          type: number
        compItemName:
          type: string
        compItemSlug:
          type: string
        similarityScore:
          type: number
        matchReason:
          type: ["string", "null"]
    WatchlistInput:
      type: object
      required: [itemId]
      properties:
        itemId:
          type: number
    WatchlistEntry:
      type: object
      required: [id, userId, itemId, createdAt]
      properties:
        id:
          type: number
        userId:
          type: number
        itemId:
          type: number
        createdAt:
          type: string
...